third_party/blink/web_tests/http/tests/inspector-protocol/preload/rule-set-updated-expected.txt


Running test: testEnabled
Tests that Preload.ruleSetUpdated and Preload.ruleSetDeleted are dispatched.
{
    _selector : #prefetch
    backendNodeId : <number>
    id : <string>
    loaderId : <string>
    sourceText : {
        prefetch : [
            [0] : {
                source : list
                urls : [
                    [0] : /subresource.js
                ]
            }
        ]
        tag : prefetch_tag
    }
    tag : prefetch_tag
}
{
    _selector : #prerender
    backendNodeId : <number>
    id : <string>
    loaderId : <string>
    sourceText : {
        prerender : [
            [0] : {
                source : list
                urls : [
                    [0] : /page.html
                ]
            }
        ]
        tag : prerender_tag
    }
    tag : prerender_tag
}
{
    _selector : #prerender-until-script
    backendNodeId : <number>
    id : <string>
    loaderId : <string>
    sourceText : {
        prerender_until_script : [
            [0] : {
                source : list
                urls : [
                    [0] : /page.html?pus=1
                ]
            }
        ]
        tag : prerender_until_script_tag
    }
    tag : prerender_until_script_tag
}
{
    _selector : #invalid-json
    backendNodeId : <number>
    errorMessage : Line: 4, column: 7, Syntax error.
    errorType : SourceIsNotJsonObject
    id : <string>
    loaderId : <string>
    sourceText : {"prefetch":[
}
{
    _selector : #not-object
    backendNodeId : <number>
    errorMessage : Parsed JSON must be an object.
    errorType : SourceIsNotJsonObject
    id : <string>
    loaderId : <string>
    sourceText : "invalid"
}
{
    _selector : #contains-invalid-rule
    backendNodeId : <number>
    errorMessage : A rule set for a key must be an array: path = ["prerender"]
    errorType : InvalidRulesSkipped
    id : <string>
    loaderId : <string>
    sourceText : {"prefetch":[{"source":"list","urls":["/subresource.js"]}],"prerender":"invalid"}
}
{
    method : Preload.ruleSetRemoved
    params : {
        id : <string>
    }
    sessionId : <string>
}

Running test: testDisabled
Tests that Preload.ruleSetUpdated and Preload.ruleSetDeleted are not dispatched.
